Major changes in translation and etc.
- Overhauled translation method to make better sense
- Image detection for feeds have been changed to account for all images
with exensions jpg, jpeg, png, gif, bmp, and webp. They are noted
according to {imageX} with X being 1-9. Usable in rssEmbed.
- Added new 'Image URL' (imageURL) property to rssEmbed. (Bottom image
of an embed). Image tags can be used in this field, as well as any other
URL field.
- Most menus should spawn on promise resolution for its triggering
message now
- Feed deletion mid-RSS process will no longer immediately delete cache
for rssServer. It will be held off until next cycle.
- When sending a message to Discord, it will now check whether the
message's feed is deleted in the updated guild profile contents (that
may be awaiting for cycle end for cache update)
- Message length/empty message will now be checked in sendToDiscord
instead of during translation
- Added catch to manypromises temporarily
- Fixed init process not working properly for feeds older than X days
- Fixed 'sending test message' message to delete properly only after
sending 2 messages
- Removed unnecessary step in feed initialization
- Cleaned up various code