Update Procfile
I got some very exciting news. Due to a slightly different way of configuring the heroku application. It's able to auto update!

So, a user forks your repo, can use backstore for himself to get a very nice pull of the orgin. Accepts that. And then the update kicks right into the heroku app!

Using the env variables makes this also secure too! The only things exposed to the public is the id of the admin.

There is only one last thing preventing this to be a fully automated, controller solution. And that is how to auto accept a pull from backstroke...