Sanitize mentions in cleanup (prior to role logic)
Per DM, opening PR. Sanitizes mentions in the raw article via. the `cleanup` function with a zero-width space. This makes all mentions (such as `@everyone` - that could be contained in the RSS body) not get triggered on Discord but does not affect the output, nor has any impact on subscribed roles, as that logic is done after this cleanup.